requirements.txt
xarray>=2026.2.0
ipykernel>=7.2.0
fastapi>=0.129.1
uvicorn[standard]>=0.41.0
Pillow>=12.1.0
pytest>=9.0.2
httpx>=0.28.1
h5netcdf>=1.8.1
h5py>=3.15.1
